started to style better. carousel appears now when more than 4 cards
